Refactor question retrieval and fetching logic into separate routers; remove old question fetcher and retriever files
